Use correct type for yyleng.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@3455 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/lexer.h
@@ -48,7 +48,7 @@ typedef token_t yylex_t(void);
 typedef struct lexer_s {
     yylex_t  *yylex;
     char    **yytext;
-    int      *yyleng;
+    size_t   *yyleng;
 } lexer_t;
 
 extern lexer_t *lexer;
